+++ src/shared/ui/pagination/dots.tsx
@@ -1,0 +1,5 @@
+import { ReturnComponent } from '@/shared/types'
+
+const Dots = (): ReturnComponent => {
+  return <span>&#8230;</span>
+}
