+++ src/shared/ui/pagination/hooks/index.ts
@@ -1,0 +1,1 @@
+export { usePagination } from './usePagination'
