+++ src/shared/ui/pagination/index.ts
@@ -1,0 +1,1 @@
+export { Pagination } from './pagination'
